backend/requirements.txt
Flask==2.2.5
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.2
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0
Flask-Cors==4.0.1
itsdangerous==2.1.2
